feat(data): connecting Firestore to add store local data to cloud and display it

+++ app/src/main/java/com/lokatani/lokafreshinventory/ui/detail/DetailViewModel.kt
@@ -17,7 +17,7 @@ class DetailViewModel(
     fun insertResult(
         user: String,
         vegResult: String,
-        vegWeight: Float,
+        vegWeight: Int,
         date: String
     ) {
         val result = ScanResult(
